fix migration problems

+++ database/migrations/2025_06_09_142259_add_order_number_to_sponsors_table.php
@@ -8,15 +8,19 @@ class AddOrderNumberToSponsorsTable extends Migration
 {
     public function up()
     {
-        Schema::table('sponsors', function (Blueprint $table) {
-            $table->unsignedInteger('order_number')->nullable()->after('sponsor_size');
-        });
+        if (Schema::hasTable('sponsors')) {
+            Schema::table('sponsors', function (Blueprint $table) {
+                $table->unsignedInteger('order_number')->nullable()->after('sponsor_size');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('sponsors', function (Blueprint $table) {
-            $table->dropColumn('order_number');
-        });
+        if (Schema::hasTable('sponsors')) {
+            Schema::table('sponsors', function (Blueprint $table) {
+                $table->dropColumn('order_number');
+            });
+        }
     }
 }
